[Refactor] #163 - 작품 검색 Domain 계약을 NovelDomain에서 SearchDomain으로 이동
