+++ Projects/Domain/NovelDomain/Testing/Mock/MockNovelRepository.swift
@@ -16,8 +16,6 @@ public final class MockNovelRepository: NovelRepository {
     public var fetchNovelResult: Result<NovelInformation, RepositoryError>!
     public var addInterestResult: Result<Void, RepositoryError> = .success(())
     public var removeInterestResult: Result<Void, RepositoryError> = .success(())
-    public var searchByTextResult: Result<(Paginated<Novel>, Int), RepositoryError>!
-    public var searchByFilterResult: Result<(Paginated<Novel>, Int), RepositoryError>!
     public var fetchMyLibraryResult: Result<(Paginated<LibraryNovel>, Int), RepositoryError>!
     public var fetchUserLibraryResult: Result<(Paginated<LibraryNovel>, Int), RepositoryError>!
     public var fetchRegisteredNovelStatsResult: Result<RegisteredNovelStats, RepositoryError>!
@@ -26,10 +24,6 @@ public final class MockNovelRepository: NovelRepository {
     public private(set) var lastCachedKeywords: [Keyword]?
     public private(set) var addedInterestIDs: [NovelID] = []
     public private(set) var removedInterestIDs: [NovelID] = []
-    public private(set) var searchByTextCallCount = 0
-    public private(set) var lastSearchQuery: String?
-    public private(set) var searchByFilterCallCount = 0
-    public private(set) var lastSearchFilter: SearchFilter?
     public private(set) var fetchedMyLibraryFilters: [MyLibraryFilter] = []
     public private(set) var fetchedUserLibraryIDs: [UserID] = []
     public private(set) var fetchedUserLibraryFilters: [LibraryFilter] = []
@@ -51,18 +45,6 @@ public final class MockNovelRepository: NovelRepository {
     public func removeNovelInterest(id: NovelID) async throws(RepositoryError) {
         removedInterestIDs.append(id)
         try removeInterestResult.get()
-    }
-
-    public func searchNovelByText(_ text: String) async throws(RepositoryError) -> (Paginated<Novel>, Int) {
-        searchByTextCallCount += 1
-        lastSearchQuery = text
-        return try searchByTextResult.get()
-    }
-
-    public func searchNovelByFilter(_ filter: SearchFilter) async throws(RepositoryError) -> (Paginated<Novel>, Int) {
-        searchByFilterCallCount += 1
-        lastSearchFilter = filter
-        return try searchByFilterResult.get()
     }
 
     public func fetchMyLibraryNovels(_ filter: MyLibraryFilter) async throws(RepositoryError) -> (Paginated<LibraryNovel>, Int) {
