+++ Projects/Domain/SearchDomain/Sources/Entity/SearchFilter.swift
@@ -1,6 +1,6 @@
 //
 //  SearchFilter.swift
-//  NovelDomain
+//  SearchDomain
 //
 //  Created by Seoyeon Choi on 2/11/26.
 //  Copyright © 2026 kr.websoso.app. All rights reserved.
