+++ Projects/Domain/SearchDomain/Sources/Repository/SearchNovelRepository.swift
@@ -1,0 +1,16 @@
+//
+//  SearchNovelRepository.swift
+//  SearchDomain
+//
+//  Created by Seoyeon Choi on 2/11/26.
+//  Copyright © 2026 kr.websoso.app. All rights reserved.
+//
+
+import Foundation
+
+import BaseDomain
+
+public protocol SearchNovelRepository {
+    func searchNovelByText(_ text: String) async throws(RepositoryError) -> (Paginated<Novel>, Int)
+    func searchNovelByFilter(_ filter: SearchFilter) async throws(RepositoryError) -> (Paginated<Novel>, Int)
+}
